Fix: Improve model predictions and service calls
This commit includes several fixes and improvements to the
model prediction pipeline and service call implementations:

- Replaced localhost with 127.0.0.1 for service endpoints.
- Added service URLs to const.py for predictor and scorer.
- Adjusted data types for outputs/variances in scorer service.
- Fixed tensor conversions and data types in feature building.
- Passed request body to predictor and scorer services.
- Updated model path and keys configuration in meta_config.json.
- Handled multiple key evictions in LSTM cache.
- Added data type to new features tensor for autoregressive rollout.
- Corrected key embedding in LSTM model.
- Converted outputs/variances to tensors in confidence
  interval calculation.
- Adjusted cache TTL and weights in the configuration file.
- Converted keys to integers for eviction decisions.

+++ src/components/dataset/features/seq_builder.py
@@ -55,12 +55,12 @@ def build_feature_seq(
         # Build times sequence and move to device
         times_seq = torch.tensor(
             np.stack([sin_time, cos_time], axis=1),
-            dtype=torch.float,
+            dtype=torch.float32,
         ).unsqueeze(0)
         times_seq = move_to_device(times_seq, device)
 
         # Build keys sequence and move to device
-        keys_seq = torch.tensor(keys, dtype=torch.long).unsqueeze(0)
+        keys_seq = torch.tensor(keys, dtype=torch.float32).unsqueeze(0)
         keys_seq = move_to_device(keys_seq, device)
 
         debug(
